Fixed turn servers fething
