+++ package.json
@@ -6,7 +6,7 @@
   "type": "module",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "clear && nodemon app2.js"
+    "start": "clear && nodemon app.js"
   },
   "keywords": [],
   "author": "",
